Moved numerical conversion functions from math.h to convert.h
