+++ test/core/mathTests.cpp
@@ -6,23 +6,6 @@
 
 namespace jactorio::core
 {
-    TEST(Math, SafeCast) {
-        auto a = SafeCast<int>(42);
-        auto c = SafeCast<uint64_t>(1); // Widening cast
-
-        // const auto b = SafeCast<float>(static_cast<double>(321321312));  // Error: Must be integral
-        // auto d = SafeCast<uint8_t>(256);  // Error: Loss of data
-    }
-
-    TEST(Math, SafeCastAssign) {
-        auto a = 42;
-        auto c = 1ull;
-
-        SafeCastAssign(a, c);
-
-        EXPECT_EQ(a, c);
-    }
-
     TEST(Math, FloatEq) {
         // Checks floats are equal within a tolerance
